+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - add "learn more"-button to manage-key section that links to local help #4684
+- add a search field to help page #4691
+- update local help (2025-02-21)
 
 ### Changed
 - open map in landscape orientation and with a bigger window #4683